refactor(notification): event playload type
