+++ openmeter/notification/event.go
@@ -2,11 +2,9 @@ package notification
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"time"
 
-	"github.com/openmeterio/openmeter/api"
 	"github.com/openmeterio/openmeter/pkg/models"
 	"github.com/openmeterio/openmeter/pkg/pagination"
 	"github.com/openmeterio/openmeter/pkg/sortx"
@@ -51,46 +49,6 @@ func (t EventType) Values() []string {
 	return []string{
 		string(EventTypeBalanceThreshold),
 	}
-}
-
-type EventPayloadMeta struct {
-	Type EventType `json:"type"`
-}
-
-func (m EventPayloadMeta) Validate() error {
-	return m.Type.Validate()
-}
-
-// EventPayload is a union type capturing payload for all EventType of Events.
-type EventPayload struct {
-	EventPayloadMeta
-
-	// Balance Threshold
-	BalanceThreshold BalanceThresholdPayload `json:"balanceThreshold"`
-}
-
-func (p EventPayload) Validate() error {
-	switch p.Type {
-	case EventTypeBalanceThreshold:
-		return p.BalanceThreshold.Validate()
-	default:
-		return ValidationError{
-			Err: fmt.Errorf("invalid event type: %s", p.Type),
-		}
-	}
-}
-
-type BalanceThresholdPayload struct {
-	Entitlement api.EntitlementMetered                    `json:"entitlement"`
-	Feature     api.Feature                               `json:"feature"`
-	Subject     api.Subject                               `json:"subject"`
-	Threshold   api.NotificationRuleBalanceThresholdValue `json:"threshold"`
-	Value       api.EntitlementValue                      `json:"value"`
-}
-
-// Validate returns an error if balance threshold payload is invalid.
-func (b BalanceThresholdPayload) Validate() error {
-	return nil
 }
 
 var _ validator = (*ListEventsInput)(nil)
@@ -329,18 +287,4 @@ func (i UpdateEventDeliveryStatusInput) Validate(_ context.Context, _ Service) e
 	}
 
 	return nil
-}
-
-func PayloadToMapInterface(t any) (map[string]interface{}, error) {
-	b, err := json.Marshal(t)
-	if err != nil {
-		return nil, err
-	}
-
-	var m map[string]interface{}
-	if err = json.Unmarshal(b, &m); err != nil {
-		return nil, err
-	}
-
-	return m, nil
 }
